Move student freq plot to vis
